Implement todos by user

+++ admin-todos/src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSession } from "@/auth/actions/auth-actions"
 import prisma from "@/lib/prisma"
 import { NextResponse, NextRequest } from "next/server"
 import * as yup from "yup"
@@ -32,12 +33,19 @@ const postSchema = yup.object({
 })
 
 export async function POST(request: Request) {
+  const user = await getUserSession()
+  if (!user) {
+    return NextResponse.json("Not authorized", { status: 401 })
+  }
+
   try {
     const { complete, description } = await postSchema.validate(
       await request.json()
     )
 
-    const todo = await prisma.todo.create({ data: { complete, description } })
+    const todo = await prisma.todo.create({
+      data: { complete, description, userId: user.id },
+    })
 
     return NextResponse.json(todo)
   } catch (err) {
@@ -46,9 +54,13 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSession()
+  if (!user) {
+    return NextResponse.json("Not authorized", { status: 401 })
+  }
   try {
     const deletedTodos = await prisma.todo.deleteMany({
-      where: { complete: true },
+      where: { complete: true, userId: user.id },
     })
     return NextResponse.json(deletedTodos)
   } catch (err) {
